# api/openapi/platform.yaml
openapi: 3.1.0
info:
  title: MyQYPT Platform API
  description: >-
    Public HTTP contract for the myqypt platform. This document is the single
    source of truth for the public API surface; transport code is generated
    from it with oapi-codegen.
  version: 0.1.0
paths:
  /api/v1/system/status:
    get:
      operationId: getSystemStatus
      summary: Report whether the platform service is available.
      responses:
        "200":
          description: The platform service is available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"
  /api/v1/tenants:
    get:
      operationId: listTenants
      summary: List the tenants the authenticated user holds an active membership in.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: >-
            Every tenant with an active membership for the authenticated user,
            with the tenant kind and the membership role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantList"
    post:
      operationId: createTenant
      summary: Create a business tenant owned by the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: >-
            Client-chosen retry key: redelivering the same key (with the same
            authenticated user) converges onto the already-created tenant.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: >-
            The business tenant was created in this delivery: the tenant, its
            display name, the creation timestamp, and the owner role the
            creator holds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedTenant"
        "200":
          description: >-
            A previous delivery with the same Idempotency-Key already created
            this tenant; the same creation facts are echoed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedTenant"
  /api/v1/tenant-context:
    get:
      operationId: getTenantContext
      summary: Return the server-validated current tenant context selection.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: >-
            The current selection, still backed by an active membership
            (revocation invalidates a persisted selection on read).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectedTenantContext"
    put:
      operationId: putTenantContext
      summary: Select or switch the current tenant context.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantContextSelection"
      responses:
        "200":
          description: >-
            The persisted selection. The request is idempotent and a switch
            is last-write-wins.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectedTenantContext"
  /api/v1/tenants/{tenantId}/membership-invitations:
    post:
      operationId: createMembershipInvitation
      summary: >-
        Invite a bound user into a tenant as a non-owner member; the
        membership activates only after the invitee accepts (T05).
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          description: The tenant the invitation grants membership in.
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          description: >-
            Client-chosen retry key, required before any write is
            attempted. Replay convergence rides the (tenant, invitee)
            natural key: a repeat delivery of a still-pending invitation
            answers 200 with the same invitation facts.
          schema:
            type: string
            pattern: "^([A-Za-z0-9-_.]{1,64})$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInvitationRequest"
      responses:
        "201":
          description: >-
            The invitation was created in this delivery: the membership row
            with status invited, invisible to the tenant-context endpoints
            until the invitee accepts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipInvitation"
        "200":
          description: >-
            A pending invitation for this (tenant, invitee) already exists;
            the same invitation facts are echoed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipInvitation"
  /api/v1/tenants/{tenantId}/membership-invitations/acceptance:
    post:
      operationId: acceptMembershipInvitation
      summary: Accept the pending membership invitation as the invitee.
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          description: The tenant whose invitation is being accepted.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: >-
            The accepted membership, now active. Replays converge onto the
            same activation without a second transition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivatedMembership"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Access token issued by the configured identity provider. Documentation
        only: the OpenAPI validator does not enforce security; authentication
        runs inside each handler.
  schemas:
    SystemStatus:
      type: object
      additionalProperties: false
      required:
        - status
        - version
      properties:
        status:
          type: string
          const: available
          description: Fixed availability marker for the platform service.
        version:
          type: string
          minLength: 1
          description: Build version of the running platform service.
    TenantSummary:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - kind
        - role
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The tenant's immutable identifier.
        kind:
          type: string
          enum:
            - personal
            - business
          description: The tenant kind (migration 000003 vocabulary).
        role:
          type: string
          enum:
            - owner
            - admin
            - billing_member
            - member
          description: The membership role the user holds in the tenant.
    TenantList:
      type: object
      additionalProperties: false
      required:
        - tenants
      properties:
        tenants:
          type: array
          items:
            $ref: "#/components/schemas/TenantSummary"
          description: Every tenant with an active membership for the user.
    TenantContextSelection:
      type: object
      additionalProperties: false
      required:
        - tenant_id
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The tenant being selected.
    SelectedTenantContext:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - selected_at
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The persisted, server-validated selected tenant.
        selected_at:
          type: string
          format: date-time
          description: When the selection was persisted or last switched.
    CreateTenantRequest:
      type: object
      additionalProperties: false
      required:
        - display_name
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
          description: The business tenant's human-readable display name.
    CreatedTenant:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - kind
        - display_name
        - created_at
        - role
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The created tenant's immutable identifier.
        kind:
          type: string
          const: business
          description: The tenant kind (T04 provisions business tenants only).
        display_name:
          type: string
          description: The created tenant's display name, echoed back.
        created_at:
          type: string
          format: date-time
          description: When the tenant was created.
        role:
          type: string
          const: owner
          description: The membership role the creator holds in the new tenant.
    MembershipRole:
      type: string
      enum:
        - admin
        - billing_member
        - member
      description: >-
        The non-owner membership role an invitation grants. owner is not
        invitable: ownership changes are out of scope for invitations.
    MembershipInvitationRequest:
      type: object
      additionalProperties: false
      required:
        - invitee_subject
        - role
      properties:
        invitee_subject:
          type: string
          minLength: 1
          maxLength: 320
          description: >-
            The invitee's external subject — the identity provider's sub
            claim of an already-bound user.
        role:
          $ref: "#/components/schemas/MembershipRole"
    MembershipInvitation:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - role
        - status
        - invited_at
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The tenant the invitation grants membership in.
        role:
          $ref: "#/components/schemas/MembershipRole"
        status:
          type: string
          const: invited
          description: >-
            The pending state: the membership row exists but is not yet
            usable at the tenant-context seam.
        invited_at:
          type: string
          format: date-time
          description: When the invitation was created.
    ActivatedMembership:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - role
        - status
      properties:
        tenant_id:
          type: string
          format: uuid
          description: The tenant of the accepted membership.
        role:
          $ref: "#/components/schemas/MembershipRole"
        status:
          type: string
          const: active
          description: The activated state the acceptance transitioned into.